segment-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
onnxruntime==1.19.2
Pillow==10.4.0
numpy==2.1.2


